Rewrote harvester using generic class to hold higher order functions
